do not rely on systemd for Debian-related families
On Debian, systemd is not mandatory, so people could still use SysV
init scripts. Moreover, the "service" commands handle both cases. On
Ubuntu, systemd is mandatory since 16.04. However, the integration
with SysV still exist and the conserver package is unmodified compared
to Debian and only ships a SysV init script.

Moreover, the override was done in /usr/lib/systemd/system, a
directory not existing on most Debian systems.

Instead, we keep acting like everyone was using SysV init and let the
appropriate compatibility bits in the distribution handle the
differences.